+++ Gears/Constants.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <cstdint>
+
+const uint8_t NUM_GEARS = 8;
+const uint8_t MAX_GEAR = 7;
+
+const char* const DISPLAY_VERSION = "v4.6.1";
+const char* const mtDir = "\\ManualTransmission";
