+++ Gears/Util/Color.h
@@ -1,0 +1,25 @@
+#pragma once
+
+struct Color {
+    int R;
+    int G;
+    int B;
+    int A;
+};
+
+const Color solidWhite = { 255,	255, 255, 255 };
+const Color solidBlack = { 0, 0, 0, 255 };
+
+const Color solidRed = { 255, 0, 0,	255 };
+const Color solidGreen = { 0, 255, 0, 255 };
+const Color solidBlue = { 0, 0, 255, 255 };
+
+const Color solidPink = { 255, 0, 255, 255 };
+const Color solidYellow = { 255, 255, 0, 255 };
+const Color solidCyan = { 0, 255, 255, 255 };
+
+const Color solidOrange = { 255, 127, 0, 255 };
+const Color solidLime = { 127, 255, 0, 255 };
+const Color solidPurple = { 127, 0, 255, 255 };
+
+const Color transparentGray = { 75, 75, 75, 75 };
