+++ Gears/Util/Files.cpp
@@ -1,0 +1,6 @@
+#include "Files.h"
+
+bool FileExists(const std::string& name) {
+    struct stat buffer;
+    return (stat(name.c_str(), &buffer) == 0);
+}
