+++ Gears/Util/Files.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <string>
+
+bool FileExists(const std::string& name);
