+++ Gears/Util/MiscEnums.h
@@ -1,0 +1,33 @@
+#pragma once
+
+//http://pastebin.com/Kj9t38KF
+enum eRadioStations {
+    LosSantosRockRadio,
+    NonStopPopFM,
+    RadioLosSantos,
+    ChannelX,
+    WestCoastTalkRadio,
+    RebelRadio,
+    SoulwaxFM,
+    EastLosFM,
+    WestCoastClassics,
+    BlueArk,
+    WorldideFM,
+    FlyLoFM,
+    TheLowdown,
+    TheLab,
+    RadioMirrorPark,
+    Space1032,
+    VinewoodBlvdRadio,
+    SelfRadio,
+    BlaineCountyRadio,
+    RadioOff = 255
+};
+
+enum eDecorType {
+    DECOR_TYPE_FLOAT = 1,
+    DECOR_TYPE_BOOL,
+    DECOR_TYPE_INT,
+    DECOR_TYPE_UNK,
+    DECOR_TYPE_TIME
+};
