+++ Gears/VehicleData.hpp
@@ -6,7 +6,7 @@
 #include <array>
 #include <vector>
 #include <chrono>
-#include "script.h"
+#include "Constants.h"
 #include "Memory/VehicleExtensions.hpp"
 
 #define SAMPLES 6
